+++ src/main/java/com/oskopek/studyguide/constraint/Constraint.java
@@ -1,8 +1,10 @@
-package com.oskopek.studyguide.constraints;
+package com.oskopek.studyguide.constraint;
 
 /**
  * A general contract for all constraints operating on the {@link com.oskopek.studyguide.model.StudyPlan} model.
  */
 public interface Constraint {
+
+
 
 }
