+++ src/main/java/com/oskopek/studyguide/constraint/CourseEnrollmentRequirementsUnfulfilledConstraint.java
@@ -1,9 +1,11 @@
-package com.oskopek.studyguide.constraints;
+package com.oskopek.studyguide.constraint;
 
 /**
  * Checks if all the required courses for a {@link com.oskopek.studyguide.model.CourseEnrollment}
- * are marked as fulfilled (at the latest this current semester).
+ * are marked as fulfilled (at the latest this current semester). Only manages and clauses.
  */
 public class CourseEnrollmentRequirementsUnfulfilledConstraint extends CourseEnrollmentConstraint {
+
+    // TODO prereqs and coreqs
 
 }
