+++ src/main/java/com/oskopek/studyguide/constraint/GlobalCourseRepeatedEnrollmentConstraint.java
@@ -1,4 +1,4 @@
-package com.oskopek.studyguide.constraints;
+package com.oskopek.studyguide.constraint;
 
 /**
  * Checks if the number of times a student is enrolled in a course is valid.
