+++ src/main/java/com/oskopek/studyguide/constraint/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Constraints the view uses to notify the user of any errors in his study plan model.
  */
-package com.oskopek.studyguide.constraints;
+package com.oskopek.studyguide.constraint;
