--- src/main/java/com/oskopek/studyguide/constraints/CourseEnrollmentConstraint.java
@@ -1,8 +1,0 @@
-package com.oskopek.studyguide.constraints;
-
-/**
- * Constraint on the level of individual {@link com.oskopek.studyguide.model.CourseEnrollment}s.
- */
-public abstract class CourseEnrollmentConstraint implements Constraint {
-
-}
