+++ src/test/java/com/oskopek/studyguide/constraint/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Tests of {@link com.oskopek.studyguide.constraint}.
+ */
+package com.oskopek.studyguide.constraint;
