--- src/test/java/com/oskopek/studyguide/constraints/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Tests of {@link com.oskopek.studyguide.constraints}.
- */
-package com.oskopek.studyguide.constraints;
